score moved to datasets module as it is a timely sorted aggregation of Abs (with some duration, albeit it was 0)
