+++ lib/musa-dsl/datasets.rb
@@ -6,3 +6,5 @@ require_relative 'datasets/packed-v'
 require_relative 'datasets/v'
 require_relative 'datasets/p'
 require_relative 'datasets/ps'
+
+require_relative 'datasets/score'
