+++ lib/musa-dsl/datasets/gdv.rb
@@ -1,4 +1,4 @@
-require_relative 'abs-d'
+require_relative 'e'
 require_relative 'gdvd'
 require_relative 'pdv'
