+++ lib/musa-dsl/datasets/pdv.rb
@@ -1,4 +1,4 @@
-require_relative 'abs-d'
+require_relative 'e'
 require_relative 'gdv'
 
 require_relative 'helper'
