--- lib/musa-dsl/score.rb
@@ -1,1 +1,0 @@
-require_relative 'score/score'
